+++ .github/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.1.9]
 ### Added
 - `-Strict` parameter to not clean filenames when scraping
+- `rename-file` setting
 
 ## [1.1.8]
 ### Fixed